Enable chapter support in the main custom player layout
