+++ Demo/Sources/Router/Router.swift
@@ -47,7 +47,7 @@ final class Router: ObservableObject {
 extension Router: PictureInPictureDelegate {
     func pictureInPictureWillStart() {
         switch presented {
-        case .player, .systemPlayer, .chaptersPlayer, .playlist, .multi:
+        case .player, .systemPlayer, .playlist, .multi:
             previousPresented = presented
             presented = nil
         case .inlineSystemPlayer:
